+++ src/Controller/Traits/CookieControllerTrait.php
@@ -1,0 +1,33 @@
+<?php
+
+
+namespace SALESmanago\Traits\Controller;
+
+
+use SALESmanago\Adapter\CookieManagerAdapter;
+
+trait CookieControllerTrait
+{
+    /**
+     * @var CookieManagerAdapter
+     */
+    protected $CookieManager;
+
+    /**
+     * @param CookieManagerAdapter $CookieManager
+     * @return $this
+     */
+    public function setCookieManager(CookieManagerAdapter $CookieManager)
+    {
+        $this->CookieManager = $CookieManager;
+        return $this;
+    }
+
+    /**
+     * @return bool
+     */
+    private function checkIfCookieAdapterSet()
+    {
+        return (isset($this->CookieManager) && $this->CookieManager != null);
+    }
+}