+++ src/Model/CreateInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace SALESmanago\Model;
-
-
-interface CreateInterface
-{
-    public function insert($settings);
-}