+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 SALESmanago library for integrations
 ------------------------------------
+Version 3.0.0 29.01.2021
+ - change structure;
+ - remove unnecessary functionality;
+ - add new mechanism for login;
+
 Version 2.6.2 20.01.2021
  - Add Event types const to Entity\Event\Event;